refactor: remove `createRecommendationsClient` from tests
