+++ packages/recommend-core/src/__tests__/getFrequentlyBoughtTogether.test.ts
@@ -1,9 +1,21 @@
-import { createRecommendationsClient } from '../../../../test/utils';
+import {
+  createMultiSearchResponse,
+  createRecommendClient,
+  hit,
+} from '../../../../test/utils';
 import { getFrequentlyBoughtTogether } from '../getFrequentlyBoughtTogether';
 
 describe('getFrequentlyBoughtTogether', () => {
   test('does not forward unwanted props', async () => {
-    const { recommendClient } = createRecommendationsClient();
+    const recommendClient = createRecommendClient({
+      getFrequentlyBoughtTogether: jest.fn(() =>
+        Promise.resolve(
+          createMultiSearchResponse({
+            hits: [hit],
+          })
+        )
+      ),
+    });
 
     await getFrequentlyBoughtTogether({
       recommendClient,
