+++ packages/recommend-core/src/__tests__/getRelatedProducts.test.ts
@@ -1,9 +1,21 @@
-import { createRecommendationsClient } from '../../../../test/utils';
+import {
+  createMultiSearchResponse,
+  createRecommendClient,
+  hit,
+} from '../../../../test/utils';
 import { getRelatedProducts } from '../getRelatedProducts';
 
 describe('getRelatedProducts', () => {
   test('does not forward unwanted props', async () => {
-    const { recommendClient } = createRecommendationsClient();
+    const recommendClient = createRecommendClient({
+      getRelatedProducts: jest.fn(() =>
+        Promise.resolve(
+          createMultiSearchResponse({
+            hits: [hit],
+          })
+        )
+      ),
+    });
 
     await getRelatedProducts({
       recommendClient,
