+++ test/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './createApiResponse';
-export * from './createRecommendationsClient';
 export * from './createRecommendClient';
